requirements.txt
redis
celery
requests
gunicorn
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-cryptography-5
# psycopg2-binary
drf-spectacular